Validation module: ValidationRule->ClientRule (internal), RuleName/RuleOperand renames, extract PlanModel/Validation/ (capability-preserving, owner-approved API)
