+++ Alis.Reactive/Validation/ValidationTerms.cs
@@ -83,77 +83,6 @@ namespace Alis.Reactive.Validation
         AllowRootField
     }
 
-    internal sealed class ValidationRuleName : IEquatable<ValidationRuleName>
-    {
-        private static readonly Dictionary<string, ValidationRuleName> Known =
-            new Dictionary<string, ValidationRuleName>(StringComparer.Ordinal)
-            {
-                { "required", new ValidationRuleName("required") },
-                { "empty", new ValidationRuleName("empty") },
-                { "minLength", new ValidationRuleName("minLength") },
-                { "maxLength", new ValidationRuleName("maxLength") },
-                { "email", new ValidationRuleName("email") },
-                { "regex", new ValidationRuleName("regex") },
-                { "url", new ValidationRuleName("url") },
-                { "creditCard", new ValidationRuleName("creditCard") },
-                { "range", new ValidationRuleName("range") },
-                { "exclusiveRange", new ValidationRuleName("exclusiveRange") },
-                { "min", new ValidationRuleName("min") },
-                { "max", new ValidationRuleName("max") },
-                { "gt", new ValidationRuleName("gt") },
-                { "lt", new ValidationRuleName("lt") },
-                { "equalTo", new ValidationRuleName("equalTo") },
-                { "notEqual", new ValidationRuleName("notEqual") },
-                { "notEqualTo", new ValidationRuleName("notEqualTo") },
-                { "atLeastOne", new ValidationRuleName("atLeastOne") },
-            };
-
-        private ValidationRuleName(string value)
-        {
-            Value = value;
-        }
-
-        internal string Value { get; }
-
-        internal static ValidationRuleName Required => Known["required"];
-        internal static ValidationRuleName Empty => Known["empty"];
-        internal static ValidationRuleName MinLength => Known["minLength"];
-        internal static ValidationRuleName MaxLength => Known["maxLength"];
-        internal static ValidationRuleName Email => Known["email"];
-        internal static ValidationRuleName Regex => Known["regex"];
-        internal static ValidationRuleName Url => Known["url"];
-        internal static ValidationRuleName CreditCard => Known["creditCard"];
-        internal static ValidationRuleName Range => Known["range"];
-        internal static ValidationRuleName ExclusiveRange => Known["exclusiveRange"];
-        internal static ValidationRuleName Min => Known["min"];
-        internal static ValidationRuleName Max => Known["max"];
-        internal static ValidationRuleName Gt => Known["gt"];
-        internal static ValidationRuleName Lt => Known["lt"];
-        internal static ValidationRuleName EqualTo => Known["equalTo"];
-        internal static ValidationRuleName NotEqual => Known["notEqual"];
-        internal static ValidationRuleName NotEqualTo => Known["notEqualTo"];
-        internal static ValidationRuleName AtLeastOne => Known["atLeastOne"];
-        internal static IReadOnlyCollection<string> Values => Known.Keys;
-
-        internal static ValidationRuleName From(string value)
-        {
-            if (value == null) throw new ArgumentNullException(nameof(value));
-            if (Known.TryGetValue(value, out var name)) return name;
-            throw new ArgumentException(
-                "Unknown validation rule '" + value + "'.",
-                nameof(value));
-        }
-
-        public bool Equals(ValidationRuleName? other) =>
-            other != null && string.Equals(Value, other.Value, StringComparison.Ordinal);
-
-        public override bool Equals(object? obj) => Equals(obj as ValidationRuleName);
-
-        public override int GetHashCode() => StringComparer.Ordinal.GetHashCode(Value);
-
-        public override string ToString() => Value;
-    }
-
     internal sealed class ValidationMessage : IEquatable<ValidationMessage>
     {
         private ValidationMessage(string value)
